fix: report the real getinfo error in check-synced; 26.6.1:1 → 26.6.1:2
check-synced parsed getinfo stdout before checking the exit code, so any
exec failure (empty stdout) surfaced as JSON.parse's "Unexpected end of
JSON input" and the intended stderr-bearing failure branch was
unreachable. Guard the exit code first and wrap the parse in try/catch,
so failures report the underlying error.

The 26.6.1:1 migration is split into its own version file (v26.6.1.1.ts)
rather than renaming current.ts in place: with a single graph node, users
updating :1 → :2 would re-run the migration and delete the regenerated
cln-grpc certs, breaking existing gRPC client pairings (e.g. Alby Hub).
Verified against the SDK version graph: :1 → :2 runs no migration; older
versions still migrate through :1.
